Split GUI native file drop tests
